PacketID Type changed to int (byte => int) 2-3 class changed with record.

+++ KavisProxy.Core/Protocol/Packets/Login/Server/Disconnect.cs
@@ -2,7 +2,7 @@
 
 public class Disconnect : Packet<Chat>
 {
-    public byte PacketID => 0;
+    public int PacketID => 0;
     public Chat Read(IByteBuffer buffer) => new Chat(buffer.ReadString());
     public void Write(IByteBuffer buffer, Chat value) => buffer.WriteString(value.RawData);
 }
