+++ KavisProxy.Core/Protocol/Packets/Status/Client/Request.cs
@@ -1,5 +1,5 @@
 ﻿namespace KavisProxy.Core.Protocol.Packets;
 public class Request : Packet
 {
-    public byte PacketID => 0;
+    public int PacketID => 0;
 }
